update javadoc and packaging and gui
