+++ src/main/java/duke/gui/Launcher.java
@@ -1,0 +1,11 @@
+package duke.gui;
+import javafx.application.Application;
+
+/**
+ * A launcher class to workaround classpath issues.
+ */
+public class Launcher {
+    public static void main(String[] args) {
+        Application.launch(Main.class, args);
+    }
+}
