+++ src/test/java/duke/EventTest.java
@@ -1,5 +1,7 @@
 package duke;
 
+import duke.exception.DukeException;
+import duke.task.Event;
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.assertEquals;
